+++ src/components/common/Modal/ChangeVoteModal/ChangeVoteModal.tsx
@@ -1,6 +1,6 @@
 import React, { SetStateAction, useEffect, useRef } from 'react';
 import Button from '@/components/common/Button/Button';
-import { useMemberQuery } from '@/hooks/api/useMemberQuery';
+import { useMemberQuery } from '@/hooks/api/member/useMemberQuery';
 import { useParseJwt } from '@/hooks/common/useParseJwt';
 import { useNewSelector } from '@/store';
 import { selectAccessToken } from '@/store/auth';
