+++ src/hooks/api/mypage/useMyVotedPostsQuery.ts
