+++ src/pages/MyPage/HistoryPage/TabPage/BookmarksPage/BookmarksPage.tsx
@@ -6,8 +6,8 @@ import List from '@/components/mypage/List/List';
 import ItemMyBookmarksPosts from '@/components/mypage/ListItem/ItemMyBookmarksPosts';
 import ItemNull from '@/components/mypage/ListItem/ItemNull';
 import { NULL } from '@/constants/message';
-import { useBookmarkDeleteAllMutation } from '@/hooks/api/useBookmarkDeleteAllMutation';
-import { useMyBookmarksPostsQuery } from '@/hooks/api/useMyBookmarksPostsQuery';
+import { useBookmarkDeleteAllMutation } from '@/hooks/api/member/useBookmarkDeleteAllMutation';
+import { useMyBookmarksPostsQuery } from '@/hooks/api/mypage/useMyBookmarksPostsQuery';
 import { useCheckboxSelect } from '@/hooks/mypage/userHistory/useCheckboxSelect';
 import { MyBookmarksPostsContentType } from '@/types/mypage';
 import { createRangeArray } from '@/utils/array';
