+++ src/pages/PostPage/CommentsSection/CommentSection/RepliesSection/RepliesSection.tsx
@@ -1,4 +1,4 @@
-import { useMemberQuery } from '@/hooks/api/useMemberQuery';
+import { useMemberQuery } from '@/hooks/api/member/useMemberQuery';
 import { useParseJwt } from '@/hooks/common/useParseJwt';
 import { useNewSelector } from '@/store';
 import { selectAccessToken } from '@/store/auth';
